requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.23.0
transformers>=4.56.0
torch>=2.1.0
torchvision>=0.16.0
accelerate==1.12.0
Pillow>=10.0.0
python-multipart>=0.0.6
pyttsx3>=2.90
kokoro
soundfile